relocate `handleRequestPayload` method
