+++ apps/web/scripts/kernel/handlers/common.ts
@@ -1,7 +1,13 @@
+import {
+	type MiniBroadcast,
+	type UnknownObject,
+	Networks,
+} from '@walless/core';
 import { modules } from '@walless/ioc';
 import { ResponseCode } from '@walless/messaging';
 import { type PublicKeyDocument, selectors } from '@walless/store';
 
+import { getRequestRecord } from '../utils/requestPool';
 import { type HandleMethod } from '../utils/types';
 
 export const handleConnect: HandleMethod = async ({
@@ -15,5 +21,20 @@ export const handleConnect: HandleMethod = async ({
 
 	responseMethod(payload.requestId, ResponseCode.SUCCESS, {
 		publicKeys: [solKey],
+	});
+};
+
+export const handleRequestPayload = (
+	payload: UnknownObject,
+	channel: MiniBroadcast,
+) => {
+	const { sourceRequestId, requestId } = payload;
+	const { payload: sourcePayload } = getRequestRecord(sourceRequestId);
+
+	return channel.postMessage({
+		...sourcePayload,
+		from: 'walless@kernel',
+		requestId,
+		responseCode: ResponseCode.SUCCESS,
 	});
 };
