+++ apps/web/scripts/kernel/handlers/kernel.ts
@@ -21,7 +21,7 @@ export const onKernelMessage: MessengerCallback = async (payload, channel) => {
 			solanaHandler.getEndpoint(payload, channel);
 			requirePrivateKey = false;
 		} else if (type === RequestType.REQUEST_PAYLOAD) {
-			coordinator.handleRequestPayload(payload, channel);
+			handleRequestPayload(payload, channel);
 		} else if (type === RequestType.SIGN_MESSAGE_ON_SOLANA) {
 			handleMethod = solanaHandler.signMessage;
 		} else if (type === RequestType.SIGN_TRANSACTION_ON_SOLANA) {
