+++ apps/web/scripts/kernel/utils/coordinator.ts
@@ -109,17 +109,3 @@ export const handle = async (
 		responseMethod: response,
 	});
 };
-
-export const handleRequestPayload = (
-	payload: UnknownObject,
-	channel: MiniBroadcast,
-) => {
-	const { sourceRequestId, requestId } = payload;
-	const { payload: sourcePayload } = getRequestRecord(sourceRequestId);
-	return channel.postMessage({
-		...sourcePayload,
-		from: 'walless@kernel',
-		requestId,
-		responseCode: ResponseCode.SUCCESS,
-	});
-};
